Add files via upload
It is a reinforcement learning (RL) project based on Super Mario using the Proximal Policy Optimization (PPO) algorithm. It's inspired by Nicholas' code, with a few fixes for running the RL training without an external display. The notebook uses 'gym_super_mario_bros' and 'nes_py' libraries for the environment.